Data Type Changes (#400)
* Data Type Changes

* Build Issue Resolved

+++ rest-api-sdk/src/main/java/com/paypal/api/payments/Cost.java
@@ -14,7 +14,7 @@ public class Cost extends PayPalModel {
 	/**
 	 * Cost in percent. Range of 0 to 100.
 	 */
-	private float percent;
+	private double percent;
 
 	/**
 	 * The cost, as an amount. Valid range is from 0 to 1,000,000.
